split render functions out
