+++ popup/src/constants.js
@@ -1,0 +1,4 @@
+const domainPrefix = "*://";
+
+export const GO_PREFIX = "go/";
+export const PREFIX = domainPrefix + GO_PREFIX;
